requirements.txt
requests
langchain
langchain_ollama
lanchain_tools
python-dotenv
